Refactor distribution analysis to support floating-point data and modularize the result UI.
- Update `EngineService`, `LocalEngineServiceImpl`, and `CalculationSystemRepository` to accept `DoubleArray` instead of `IntArray` for distribution calculations.
- Update `DistributionParams` and `MainScreenEvent` to use `List<Double>` for input data.
- Modify `DistributionForm` logic to parse input sequences as doubles rather than integers.
- Extract distribution result layout from `ResultForm.kt` into a standalone `DistributionResultScreen` component.
- Enhance `SingleResultCard` and `ComplexResultCard` with improved styling, including inner shadows, rounded corners, and consistent background colors.
- Add `fillMaxSize` modifier to `CustomEditText` for better layout consistency.

+++ app/src/main/java/com/example/stochia/core/services/calculation_system/interface/EngineService.kt
@@ -8,5 +8,5 @@ interface EngineService {
 
     fun genMarkov(states: IntArray, probs: DoubleArray, initState: Int, steps: Int): PyObject
 
-    fun getDistribution(params: IntArray): PyObject
+    fun getDistribution(params: DoubleArray): PyObject
 }
